msvc / windows related fixes courtesy of tobigun
git-svn-id: https://projectm.svn.sourceforge.net/svnroot/projectm/trunk@1330 6778bc44-b910-0410-a7a0-be141de4315d

+++ src/libprojectM/Renderer/PerlinNoise.cpp
@@ -40,7 +40,7 @@ PerlinNoise::PerlinNoise()
 	int seed = rand()%1000;
 
 	int size = 512;
-	int octaves = sqrt(size);
+	int octaves = sqrt((double)size);
 
 		for (int x = 0; x < size;x++)
 			for (int y = 0; y < size;y++)
